+++ cougar/cext/rolling_var.h
@@ -7,10 +7,10 @@
 
 #define Method var
 
-#define Rolling_Signature(name, dtype)                                           \
-    static void Rolling_Concat(rolling_##name, dtype)(PyArrayObject * source,    \
-                                                      PyArrayObject * target,    \
-                                                      int window, int min_count, \
+#define Rolling_Signature(name, dtype)                                                 \
+    static void Rolling_Concat(rolling_##name, dtype)(PyArrayObject * source,          \
+                                                      PyArrayObject * target,          \
+                                                      size_t window, size_t min_count, \
                                                       int axis, int ddof)
 
 #define Rolling_Init() \
